Fixed exception for begin and end points of repair railway event

+++ Models/RepairRailwayEvent.cs
@@ -4,7 +4,7 @@
     {
         public int ID { get; set; }
         public int RegionID { get; set; }
-        public double? BeginPoint { get; set; }
-        public double? EndPoint { get; set; }
+        public int? BeginPoint { get; set; }
+        public int? EndPoint { get; set; }
     }
 }
